style: reset css 소스 index.css로 이동 및 기존 소스 삭제
